# src/schemas/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI-PLATFORM Intelligent Automation Platform API
  description: |
    Comprehensive API specification for the AI-PLATFORM platform, covering all microservices
    and their endpoints for workflow orchestration, content generation, and system management.

    ## Architecture
    - **API Gateway**: Central entry point with authentication and rate limiting
    - **Intent Parser**: Natural language processing and workflow planning
    - **Workflow Orchestrator**: Temporal.io-based workflow execution
    - **Progress Monitor**: Real-time workflow tracking and notifications
    - **Analytics Service**: Business intelligence and performance metrics
    - **Federation Service**: External client integration via MCP protocol

    ## Authentication
    All endpoints require JWT bearer token authentication unless specified otherwise.
    Tokens are obtained via the `/auth/login` endpoint and must be included in the
    Authorization header: `Authorization: Bearer <token>`

    ## Rate Limiting
    - **Free Tier**: 60 requests/minute, 1,000 requests/hour
    - **Pro Tier**: 1,000 requests/minute, 10,000 requests/hour
    - **Enterprise**: Custom limits based on agreement

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Request limit per window
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Reset`: Unix timestamp when window resets

  version: 1.0.0
  contact:
...
  license:
    name: MIT
...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User authentication
      description: Authenticate user with email/password and return access tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Get new access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
        "401":
          description: Invalid refresh token
  /auth/logout:
    post:
      tags: [Authentication]
      summary: User logout
      description: Invalidate current session and tokens
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /workflows:
    get:
      tags: [Workflows]
      summary: List workflows
      description: Get paginated list of user workflows
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/WorkflowType"
        - $ref: "#/components/parameters/WorkflowStatus"
        - $ref: "#/components/parameters/SortBy"
        - $ref: "#/components/parameters/SortOrder"
      responses:
        "200":
          description: Workflows retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListWorkflowsResponse"

    post:
      tags: [Workflows]
      summary: Create workflow definition
      description: Create new workflow definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowRequest"
      responses:
        "201":
          description: Workflow created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateWorkflowResponse"
  /workflows/{workflowId}:
    get:
      tags: [Workflows]
      summary: Get workflow details
      parameters:
        - $ref: "#/components/parameters/WorkflowId"
      responses:
        "200":
          description: Workflow details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDefinition"

    put:
      tags: [Workflows]
      summary: Update workflow definition
      parameters:
        - $ref: "#/components/parameters/WorkflowId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkflowRequest"
      responses:
        "200":
          description: Workflow updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDefinition"

    delete:
      tags: [Workflows]
      summary: Delete workflow definition
      parameters:
        - $ref: "#/components/parameters/WorkflowId"
      responses:
        "204":
          description: Workflow deleted
  /workflows/{workflowId}/execute:
    post:
      tags: [Workflows]
      summary: Execute workflow
      description: Start execution of workflow definition
      parameters:
        - $ref: "#/components/parameters/WorkflowId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteWorkflowRequest"
      responses:
        "201":
          description: Workflow execution started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecuteWorkflowResponse"

  # =====================================================================
  # Progress Monitoring
  # =====================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token obtained from login endpoint
  parameters:
    WorkflowId:
      name: workflowId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique workflow identifier
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
    WorkflowType:
      name: workflow_type
      in: query
      schema:
        $ref: "#/components/schemas/WorkflowType"
      description: Filter by workflow type
    WorkflowStatus:
      name: status
      in: query
      schema:
        $ref: "#/components/schemas/WorkflowExecutionStatus"
      description: Filter by workflow status
    SortBy:
      name: sort_by
      in: query
      schema:
        type: string
        enum: [created_at, updated_at, name, status]
        default: created_at
      description: Sort field
    SortOrder:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      description: Sort order
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        email_verified:
          type: boolean
        status:
          $ref: "#/components/schemas/UserStatus"
        subscription_tier:
          $ref: "#/components/schemas/SubscriptionTier"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
          nullable: true
    WorkflowDefinition:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        workflow_type:
          $ref: "#/components/schemas/WorkflowType"
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStep"
        status:
          $ref: "#/components/schemas/WorkflowDefinitionStatus"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WorkflowExecution:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workflow_definition_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        temporal_workflow_id:
          type: string
        status:
          $ref: "#/components/schemas/WorkflowExecutionStatus"
        progress_percent:
          type: number
          minimum: 0
          maximum: 100
        current_step:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        estimated_completion:
          type: string
          format: date-time
          nullable: true
        total_cost_usd:
          type: number
          minimum: 0
        error_message:
          type: string
          nullable: true
    WorkflowStep:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        step_type:
          $ref: "#/components/schemas/StepType"
        function_name:
          type: string
        parameters:
          type: object
        dependencies:
          type: array
          items:
            type: string
            format: uuid
        retry_config:
          $ref: "#/components/schemas/RetryConfig"
        timeout_seconds:
          type: integer
          nullable: true
        order:
          type: integer
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        remember_me:
          type: boolean
          default: false
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        token_type:
          type: string
          default: Bearer
        user:
          $ref: "#/components/schemas/User"
...
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    RefreshTokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
    LogoutRequest:
      type: object
      properties:
        session_token:
          type: string
        logout_all_sessions:
          type: boolean
          default: false
...
    CreateWorkflowRequest:
      type: object
      required: [name, workflow_type, steps]
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        workflow_type:
          $ref: "#/components/schemas/WorkflowType"
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStepRequest"
    WorkflowStepRequest:
      type: object
      required: [name, step_type, function_name, parameters, order]
      properties:
        name:
          type: string
        step_type:
          $ref: "#/components/schemas/StepType"
        function_name:
          type: string
        parameters:
          type: object
        dependencies:
          type: array
          items:
            type: string
            format: uuid
        retry_config:
          $ref: "#/components/schemas/RetryConfig"
        timeout_seconds:
          type: integer
        order:
          type: integer
    CreateWorkflowResponse:
      type: object
      properties:
        workflow:
          $ref: "#/components/schemas/WorkflowDefinition"
        message:
          type: string
    UpdateWorkflowRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStepRequest"
        status:
          $ref: "#/components/schemas/WorkflowDefinitionStatus"
    ExecuteWorkflowRequest:
      type: object
      properties:
        input_parameters:
          type: object
        priority:
          $ref: "#/components/schemas/WorkflowPriority"
        scheduled_at:
          type: string
          format: date-time
        budget_limit_usd:
          type: number
    ExecuteWorkflowResponse:
      type: object
      properties:
        execution:
          $ref: "#/components/schemas/WorkflowExecution"
        message:
          type: string
    ListWorkflowsResponse:
      type: object
      properties:
        workflows:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowDefinition"
        total_count:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        has_more:
          type: boolean
...
    RetryConfig:
      type: object
      properties:
        max_attempts:
          type: integer
          default: 3
        initial_delay_ms:
          type: integer
          default: 1000
        max_delay_ms:
          type: integer
          default: 30000
        backoff_multiplier:
          type: number
          default: 2.0
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ApiError"
        success:
          type: boolean
          default: false
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
...
    UserStatus:
      type: string
      enum: [active, suspended, deleted]
    SubscriptionTier:
      type: string
      enum: [free, pro, enterprise]
    WorkflowType:
      type: string
      enum:
        [
          content_creation,
          marketing_campaign,
          scheduled_publishing,
          ecommerce_automation,
          web_development,
          analytics,
          hybrid,
          custom,
        ]
    WorkflowDefinitionStatus:
      type: string
      enum: [draft, active, paused, archived]
    WorkflowExecutionStatus:
      type: string
      enum: [queued, running, paused, completed, failed, cancelled, timed_out]
    StepType:
      type: string
      enum:
        [
          content_creation,
          marketing_campaign,
          scheduling_publishing,
          ecommerce_operation,
          client_integration,
          analytics,
          notification,
          data_transformation,
        ]
...
    WorkflowPriority:
      type: string
      enum: [low, normal, high, urgent]
...